Extract configureMetadata to ./configure.ts.
