+++ packages/orm/src/relations/PolymorphicReference.ts
@@ -1,6 +1,7 @@
 import { Entity, isEntity } from "../Entity";
 import { IdOf, TaggedId, sameEntity } from "../EntityManager";
 import { PolymorphicFieldComponent, getMetadata } from "../EntityMetadata";
+import { maybeGetConstructorFromReference } from "../configure";
 import { getField, setField } from "../fields";
 import {
   OneToOneReference,
@@ -10,7 +11,6 @@ import {
   fail,
   getConstructorFromTaggedId,
   getInstanceData,
-  maybeGetConstructorFromReference,
   maybeResolveReferenceToId,
 } from "../index";
 import { AbstractRelationImpl } from "./AbstractRelationImpl";
